git-svn-id: https://gsphere.genoscope.cns.fr/svn/NGL/trunk/ngl@3985 d0e871a8-0b98-4e59-9d77-5ea3c331444f

+++ lib-ngl-common/app/models/laboratory/container/instance/Content.java
@@ -29,7 +29,7 @@ public class Content implements IValidation{
 	// Reference Project code
 	public String projectCode;
 	
-	public Float percentage;
+	public Double percentage;
 	
 	public Map<String,PropertyValue> properties;
 	public Content(){
